add Bech32-encoded address support
introduce CBase58Address and CBech32Address
make CBitcoinAddress a wrapper that returns CBase58Address or
CBech32Address depending on input data

introduce to_redeemScript() method,
because for P2WPKHBitcoinAddress redeem script is not the same
as scriptPubKey.

unfortunately, for P2WSHBitcoinAddress to_redeemScript() cannot
be implemented, as we need all pubkeys for that,
and not only the hash.